requirements.txt
python-telegram-bot[webhooks]==21.9
aiohttp
python-dotenv
flask
gunicorn